% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BirdColours.R
\name{bird_palette}
\alias{bird_palette}
\title{A palette generator based on BIRDS}
\usage{
bird_palette(name, n, type = c("discrete", "continuous"))
}
\arguments{
\item{name}{Name of desired palette. Choices are:
\code{Kingfisher}, \code{BlueTit},  \code{ChafFinch},
\code{GoldFinch}, \code{GreatTit},  \code{HerringGull}, \code{Kestrel},
\code{Mallard},  \code{Mandarin} , \code{Parakeet} ,
\code{Pheasant}, \code{Robin}, \code{Shoveller}, \code{Sparrowhawk},
\code{WoodPigeon}}

\item{n}{Number of colors desired. Most palettes have
4 or 5 colors.
If omitted, uses all colours.}

\item{type}{Either "continuous" or "discrete". Use continuous if you want
to automatically interpolate between colours.
@importFrom graphics rgb rect par image text}
}
\value{
A vector of colours.
}
\description{
These are a handful of color palettes from photos of birds.
}
\examples{
bird_palette("WoodPigeon")
bird_palette("Mandarin")
bird_palette("Parakeet")
bird_palette("HerringGull", 3)

# If you need more colours than normally found in a palette, you
# can use a continuous palette to interpolate between existing
# colours
pal <- wes_palette(21, name = "HerringGull", type = "continuous")
image(volcano, col = pal)
}
\keyword{colors}
