% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BirdColours.R
\docType{data}
\name{bird_palettes}
\alias{bird_palettes}
\title{Complete list of palettes}
\format{
An object of class \code{list} of length 15.
}
\usage{
bird_palettes
}
\description{
Use \code{\link{wes_palette}} to construct palettes of desired length.
}
\keyword{datasets}
